restructure and get quote progress
