+++ lumibot/tools/__init__.py
@@ -29,4 +29,3 @@ from .pandas import *
 from .types import *
 from .yahoo_helper import YahooHelper
 from .ccxt_data_store import CcxtCacheDB
-from .ib_client_portal import IBClientPortal
